# Conflicts:
#	src/main/kotlin/hu/bme/customerqueueappbackend/contoller/CustomerTicketController.kt
#	src/main/kotlin/hu/bme/customerqueueappbackend/dto/CustomerServiceDto.kt
#	src/main/kotlin/hu/bme/customerqueueappbackend/service/CustomerServiceServiceImpl.kt

+++ src/main/kotlin/hu/bme/customerqueueappbackend/service/ServiceTypeServiceImpl.kt
@@ -3,30 +3,34 @@ package hu.bme.customerqueueappbackend.service
 import hu.bme.customerqueueappbackend.dto.ServiceTypeDto
 import hu.bme.customerqueueappbackend.dto.request.CreateServiceTypeRequest
 import hu.bme.customerqueueappbackend.model.ServiceType
+import hu.bme.customerqueueappbackend.repository.CustomerServiceRepository
+import hu.bme.customerqueueappbackend.repository.CustomerTicketRepository
 import hu.bme.customerqueueappbackend.repository.ServiceTypeRepository
+import hu.bme.customerqueueappbackend.util.exceptions.BadRequestException
+import hu.bme.customerqueueappbackend.util.exceptions.EntityNotFoundException
 import hu.bme.customerqueueappbackend.util.extensions.toDto
 import org.modelmapper.ModelMapper
 import org.springframework.data.repository.findByIdOrNull
-import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
-import org.springframework.web.server.ResponseStatusException
 import java.util.*
 import javax.transaction.Transactional
 
 @Service
 class ServiceTypeServiceImpl (
     private val serviceTypeRepository: ServiceTypeRepository,
+    private val customerServiceRepository: CustomerServiceRepository,
+    private val customerTicketRepository: CustomerTicketRepository,
     private val mapper: ModelMapper
 ): ServiceTypeService {
 
     @Transactional
     override fun saveServiceType(createServiceTypeRequest: CreateServiceTypeRequest): ServiceTypeDto {
-        return ServiceTypeDto().run {
-            val serviceType = ServiceType(name = createServiceTypeRequest.name, handleTime = createServiceTypeRequest.handleTime)
-            serviceTypeRepository.save(serviceType)
-            // Todo: add service type to customer service by createServiceTypeRequest.customerServiceId
-            serviceType.toDto(mapper)
-        }
+        val customerService = customerServiceRepository.findByIdOrNull(createServiceTypeRequest.customerServiceId) ?: throw EntityNotFoundException("Customer service was not found")
+        val serviceType = ServiceType(name = createServiceTypeRequest.name, handleTime = createServiceTypeRequest.handleTime, customerService = customerService)
+        
+        serviceTypeRepository.save(serviceType)
+
+        return serviceType.toDto(mapper)
     }
 
     override fun getServiceTypes(customerServiceId: UUID): List<ServiceTypeDto> {
@@ -36,12 +40,12 @@ class ServiceTypeServiceImpl (
 
     @Transactional
     override fun deleteServiceType(id: UUID) {
-        val serviceType = findServiceTypeById(id)
-        serviceTypeRepository.delete(serviceType)
-    }
+        val serviceType = serviceTypeRepository.findByIdOrNull(id) ?: throw EntityNotFoundException("Service Type not found")
 
-    private fun findServiceTypeById(id: UUID): ServiceType {
-        return serviceTypeRepository.findByIdOrNull(id)
-            ?: throw ResponseStatusException(HttpStatus.NOT_FOUND, "Service Type not found")
+        if (customerTicketRepository.findFirstByServiceType(serviceType) == null) {
+            serviceTypeRepository.delete(serviceType)
+        } else {
+            throw BadRequestException("At least one ticket has not been completed in this service type")
+        }
     }
 }
